+++ roles/speedtest/templates/container/entrypoint.sh
@@ -1,22 +1,31 @@
 #!/bin/sh
- 
+
 stop()
 {
-    echo "Shutting iperf3 service"
+    echo "Shutting down iperf3 service"
     PID="$(pidof iperf3)"
-    kill -s TERM $PID
-
-    echo "Shutting down system service"
+    if [ "$PID" != "" ]
+    then
+        kill -s TERM $PID
+    fi
+    echo "Shutting down nodejs service"
     PID="$(pidof node)"
-    kill -s TERM $PID
+    if [ "$PID" != "" ]
+    then
+        kill -s TERM $PID
+    fi
     wait
     exit
 }
 
 start()
 {
+    # no need to wait for iperf, because we start process directly
+    #ls -al /dev/stdout && iperf3 -s >> /proc/self/fd/1 2>&1 &
+    iperf3 --server --forceflush &
+
     docker-entrypoint.sh node server.js &
-    iperf3 -s &
+    COUNTER=0; while ! pidof node > /dev/null && [ "$COUNTER" != "10" ]; do sleep 0.5 && COUNTER=$((COUNTER + 1)); done
 }
 
 start
@@ -26,5 +35,6 @@ trap "stop" SIGTERM SIGINT
 # wait forever or until we get SIGTERM or SIGINT
 #while :; do sleep 360 & wait; done
 while pidof node > /dev/null && pidof iperf3 > /dev/null; do sleep 5 & wait; done
+#while pidof node > /dev/null; do sleep 5 & wait; done
 
 exit 1
